WIP: Gerrit: Reconfigure tenant when branch is directly updated
Directly pushing to git, either fast-forward or force-push is
unusual in a gerrit environment but can happen.

When a branch is directly updated (oldrev and newrev different
from 0) whether it's a fast-forward or a forced-push modification,
the same actions as branch deletion or creation should be triggered
as it can also change zuul configuration.

'ref-updated' events are thrown when zuul merges a change
in parallel of 'change-merged' event, so those are ignored here, as
'change-merged' event already triggers a event.branch_updated event

In that case a Branch(Ref) object returned as a change has an empty,
model is updated to handle this case

NOTE: ideal solution is described line 169, but as it can already
happen, we should trigger reconfiguration as we cannot (easily)
get modified files

Change-Id: I5a95b5317b0cba836d7daf601c91ae2fce4f982a