Gerrit: Fix branch name for ref-updated event
changes are using "data.change.branch" to fill event.branch
which doesn't have 'refs/heads/' prefix

branch creation and deletion were using "data.refUpdate.refName"
to fill event.branch which has the complete refname: it starts
with 'refs/heads/'

As a result, in the Scheduler event process queue, the cache was not
properly cleared when calling class Abide's clearUnparsedBranchCache,
called using event.branch and not change.branch, which is correctly
set to the relative refname

When reconfiguring a tenant, it could hold data which were removed
and reconfigured tenant could still use data that had disappeared

This patch removes the prefix "refs/heads/" when setting event branch
name in case of deletion or creation, change branch name is retrieve
from event.

note: it's not possible to use event.branch to set change.branch in
gerritconnection getChange as event object has not a constant base
type.
it can be either:
- GerritTriggerEvent -> TriggerEvent -> object
- DequeueEvent -> ManagementEvent -> object
the latter doesn't have a branch attribute

Change-Id: Ie60b382b23074cc9feff0648e786ddaf0d3454aa