Fix branch name and project name for ref-updated create/delete
1.
gerritconnection: changes are using "data.change.branch" to fill
event.branch which doesn't have 'refs/heads/' prefix

branch creation and deletion were using "data.refUpdate.refName"
to fill event.branch which has the complete refname: it starts
with 'refs/heads/'

As a result, in the Scheduler event process queue, the cache was not
properly cleared when calling Abide's clearUnparsedBranchCache,
called using event.branch and not change.branch.

change.branch is already correctly set to the relative refname

When reconfiguring a tenant, it could hold data which were removed
and reconfigured tenant could still use data that had disappeared

This patch removes the prefix "refs/heads/" when setting event branch
name in case of deletion or creation

note: it's not possible to use event.branch to set change.branch in
gerritconnection.getChange as event objects doesn't have a constant base
type.
it can be either:
- GerritTriggerEvent -> TriggerEvent -> object
- DequeueEvent -> ManagementEvent -> object
the latter doesn't have a branch attribute

2.
Setting event.branch is moved to any 'ref-update' event, so it will now
be set when newrev and oldrev are both not null. This doesn't have and
should not have any impacts

Change-Id: Ie60b382b23074cc9feff0648e786ddaf0d3454aa