Correctly handle skipped jobs on retries
There is a scenario where skipping child jobs combined with retries
can lead to hanging buildsets. The only way to recover them is a
dequeue and re-enqueue wither by close/reopen the change or using the
cli for non-change based pipelines.

1. cache pauses
2. skip-upload skips all child jobs
3. test does a retry which resets upload which must get skipped
   again during the reset process because of skip-upload skipping it.

cache (pauses) -+
                |
                |
                +--> test (retries) -----------+
                                               |
                                               +--> upload (skipped)
                                               |
                +--> prepare-upload (skipped) -+
                |
skip-upload ----+

Currently the job upload in this scenario gets reset to none and
indefinitely waits for being scheduled.

This can be fixed by calling setResult on all finished jobs similar to
what is done during a reenqueue. This will reset the build result to
SKIPPED again if skipping is still the right choice.

Change-Id: Ia51e46f71267535fef792d875b271271305fdad2