FakeGerritChange: Add Change-Id in commit message
Change-Id is a mandatory element for a gerrit change, it should
be present in the commit message during tests, as some queries
can look for this value in the commit message and not just rely
on metadata.

-In gerritconnection _updateChangeDependencies
-When checking NeededByFromCommit
-In FakeGerritConnection _simpleQuery, the condition:
"if msg in change.data['commitMessage']" is now valid

The following query:
"message:{I7dc1d8fe1dd8d34ac99d981517bc3879750445da}"
now finds a change with has the following message:
"A\nChange-Id:I7dc1d8fe1dd8d34ac99d981517bc3879750445da"
So an additional change query is performed

As a result:
test_inventory is updated to compare against, and to compute the
base64 using the commit message.

TestScheduler.test_source_cache: change.query() is called more
often, queried counters are increased

Change-Id: Ia65134aa1956ce1da3e0ba21c4860bc7dfe438af