PF4: Update build result page layout
Following up on the patterns used for the buildset result page this now
updates the layout of the build result page in the same manner.

Additionally, the old "Summary" part is now always shown above the tabs
(Logs, Console, ...) and the first tab was changed to show the build
results.

The Logs and Console tabs are unchanged.

Change-Id: If04b65596f189c343cda05a4a8fb3e1b740a0be2