WIP: keep retrying SQL db init
If we can't connect to the database at startup, keep trying to do
so each time Zuul tries to access the database.

This should make the scheduler more self-healing in case the db is
not available on start.  However, it could also mean that the first
report after such a situation may take a long time if there is a
large schema change pending.

Change-Id: I38ba760a884c879d5f6da6137450669c3ec4f9ae