PF4: Add icons and empty state to artifacts section on build result page
This adds some icons ot the artifact list on the build result page
showing the type of the artifact. For the assignment of the icon I'm
using a simple mapping approach, which might not be the right way on a
long term as the type could be anything specified by the user.

Additionally the patch shows an empty state if the build does not
provide any artifacts.

For the TreeView I did not find any equivalent component in PF4, so we
have to think about how to adapt this or implement a TreeView component
by our own.

Change-Id: Id8347cffcda58237760025dcfa3d0bd5ce36f4e4