scheduler: Fix event process abide hasUnparsedBranchCache argument
hasUnparsedBranchCache 1st argument is expecting the project's canonical name.
This could lead to cache not being properly cleaned

As a result 2 tests needed a fix

1. a github test was no longer working. In the current test class, the tenant
is not configured to globaly or per project exclude unprotected branches, so a
reconfigure is to expect.

A copy of this test is added in TestGithubUnprotectedBranches to use a project
which has exclude-unprotected-branches enabled.

Various fixes are also included:
- getPushEvent: no mutable default argument [1]
- branch_updated only when required
- change test structure of similar tests to send the proper sha1 when deleting
  the branch

2. an exception appeared in test_pending_merge_in_reconfig:

Traceback (most recent call last):
  File "/home/safranaerosystemsfcs/zuul/zuul/zuul/scheduler.py", line 1347, in process_management_queue
    self._doTenantReconfigureEvent(event)
  File "/home/safranaerosystemsfcs/zuul/zuul/zuul/scheduler.py", line 911, in _doTenantReconfigureEvent
    project.canonical_name, branch)
AttributeError: 'NoneType' object has no attribute 'canonical_name'

after setting the project, the test was blocked during tenant reconfiguration.
It needs a merger to be completed.
This change reactivates the merger before starting a reconfigure.
Some additional checks are making sure the initial merge job did not complete

[1] https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments

Change-Id: I1e0948944f71a74411e78e3b3e6b7eaf2ed36b63