Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

It seems like on multi tenant dashboard, this is failing with `Network Error` as noted in the comment inlined in Status.jsx.

Also is it expected that the refresh button are no longer available in the different page?

Finally I really like the new "fetching info" message!

Patch-set: 7
Label: Code-Review=-1