Add an icon next to result buttons on the console log
As a low-cost attempt to indicate that the results button is
clickable and will bring up details, add a search-with-plus icon
("zoom in") to the left of it.  Let's see if that indicates to
users that they can click on it (or at least mouseover and read
the help text that explains it).

Change-Id: I3fff295d2b3b38b9e1c75e13f9a8a967e812168e