Update patch set 7
Patch Set 7:

(1 comment)

I've removed the refresh button from some pages where it doesn't make much sense IMHO (see commit message). E.g. the build and buildset result pages are static pages (once the data is stored in the database it doesn't change). Thus, clicking on "refresh" always reloads the same content (correct me if I'm wrong here). On the other hand, Change and Status are really dynamic pages that change over time.

We could still add the Fetchable component to other pages as well if we want. I just think that a "refresh" indicates to the user that something can change and new content will be shown (otherwise you could always fall back to the browser refresh).

While writing that I realise that I also removed the refresh from some "list" pages - like projects, nodes and labels. For those, it makes sense so I will add it to those again. It might also make sense to add it to other list pages like builds and buildsets where the refresh was missing completely.

With that in mind - only the "single" pages like Project, Build, Buildset and Job would be left, so we could argue if we add the refreshable also to those - so it is consistent across all pages.

Patch-set: 7
Label: Verified=0