Update patch set 3
Patch Set 3: Code-Review-1

Generally I like the idea of a settings page.

I noticed a few problems with this:

* auto-reload is off by default (should be on).
* With off by default the status page doesn't show without clicking refresh.
* Enabling auto-reload doesn't trigger the reloads and requires a manual reload or switching browser tabs

A general note, I don't know if that's already done or discussed but we should align this work with the login work. In order to not run into conflicts I think we should either rebase the login work onto this or the other way round before merging either of those so we can be sure that this works together correctly.

Patch-set: 3
Label: Code-Review=-1