Fix executor stats with zoned executors
Currently the executor stats (accepting, queue, running) are taken
from the last executor zone encountered in the variants of the
'executor:execute.*' functions. When having zoned executors this is
wrong and instead we should sum up those values. Also add a todo to
add per zone stats in the future.

Change-Id: I1c5f000bc34921782e99718f73978b4de9a166e1