Consolidate summary, logs and console on a single build page
So far, these three tabs were split up in three different Pages to allow
each page being reachable via a different URL (rather than a #<tab>
anchor which is usually used for tabs). This makes changing the builds
page much more difficult as we have keep all three files in sync.

React-router allows a Route to pass custom props to the component that
is rendered when the Route matches a certain URL. We can use this
functionality to provide the active tab via a property for each route
using to the BuildPage component.

This change keeps the way the tabs are working, but maps all three URLs
to a single component with different props to show the activate tab.

Change-Id: Icb5729d759bafdfc396ea3642c8c6c71b3461d00