Update patch set 11
Patch Set 11:

Regarding "table vs list" (if that was your point): Do you think we need the table headers (field names) to understand which fields are shown?

I think a data list provides more flexibility in layout and styling in case we want to add more attributes to that list in the future. I'm also planning to use a similar list (just with more information) for the /builds and /buildsets pages and with that it's - in my opinion - more convenient to have both "build lists" styled in a similar fashion ;-)

AFAIK data lists are also searchable, sortable and filterable like tables. Just the sorting might look different as we don't have column headers were we can place the sort icon beside.

Patch-set: 11