Update patch set 11
Patch Set 11:

Sorry, forgot to reply to the cover message (re table vs list):

I think in this one case, the headers aren't necessary.  But I think they're going to be a big help almost everywhere else where we have more columns (eg, the list of builds and buildsets).  I also think that having a place to put sort buttons will be useful too.

I don't object to the list, but I have my doubts it's going to work for the more complex pages.  For now, my opinion is we should use a table, but I'm open to changing my mind if you want to mock up the top-level build list table (that one has the most columns).

Patch-set: 11