Update patch set 11
Patch Set 11:

Thanks for the response. I think I got your point :)
I'd propose the following: We keep the list here for now as I think that a table adds too much overhead which is not necessary here. We don't need  table headers as all relevant attributes are already "extracted" in the first part (buildset summary) of the page (project, pipeline, branch, ...). We also don't need any sorting or filtering here since the number of builds in a buildset should be manageable. In case we want that we would still be able to add a simple toolbar above the list e.g. sort builds by name or time (everything else won't make much sense IMHO). So, just a small list to keep things clean and simple ;-)

For the builds and buildsets pages a table might come in handy due to the number of attributes shown and to easier incorporate a toolbar for searching and filtering. With the new layout and classes we should still be able to get some nice looking tables.

That would also fulfil your point that the pages could look different so a user knows which page he is looking at ;-)

Would that be ok for you?

Patch-set: 11
Label: Verified=0