PF4: Update filter toolbar on builds and buildsets page
This updates the toolbar on the builds and buildsets pages with new PF4
components. The toolbar itself doesn't handle the state of the filters,
but it provides some methods to simplify the state handling in the
parent component. This is necessary, since the filters are also used in
other parts of the page and thus must be handled in the page component
itself.

Change-Id: Iec93c1be01ec1ac096e8d3dcfa501f3488f8a3bf