Update patch set 8
Patch Set 8: Code-Review+2

I really like the direction, and happy to merge this as-is, but I think a few things could be improved:

* Maybe a little less vertical space above the tabs?
* Maybe put artifacts in their own tabs?

One thing that we want to keep in mind is that we guess where the failure is and try to display it to the user immediately when they visit the page.  With extra vertical space and artifacts, we may end up pushing that below the bottom of the screen and they may never see it.  Then they're likely to click on the logs or console tab to try to find the error.  So ensuring that it's visible in a typical situation is important.

Patch-set: 8
Label: Code-Review=+2