Update tests/base.py to use proper git data
The purpose of this change is to prepare for following tenant
reconfiguration changes, they also apply to a large variety of
tests which require small modifications. It will improve the visibility
of next changes.

addFakeChangeToRepo: when creating a new change for a project/branch
the most common starting point is the current project/branch state.
using an older one when creating a change is still possible with"parent"
attribute.

setMerged/getChangeMergedEvent/getRefupdatedEvent: when stream-events
is used, "change-merged" event is sending newRev as the new revision
of the branch, which may not be equal to the latest "revision" of the
change. "ref-updated" is linked to a branch, it has the proper newRev.
Both events should only be thrown if the associated change has been
merged into the repository.

Fix tests to merge change before getting sha1 and sending events.

Change-Id: I7f1730d6d06ab317b448e3e081dc01729131681b