Tenant reconfiguration: Scheduler ref-updated create/delete
branch_deleted reconfiguration
-----------------------------
During TenantParser fromYaml:
- on initial loading
- on reconfiguration
for every (tenant,project,branch) tuple a cache entry is created
(see configloader.py:1772 _cacheTenantYAML abide.getUnparsedBranchCache)

So, on branch_deleted event, hasUnparsedBranchCache always return True, which
triggers a reconfiguration, even if no config files are present

hasUsefulBranchCache is introduced to return True only if the tenant/project/branch
has valid zuul configuration elements, which is used to trigger or not a tenant
reconfiguration

branch_created reconfiguration
------------------------------
When a branch is created, no cache entry exists for that branch, a
reconfiguration could only start if the associated commit has valid
zuul configuration elements

for a Change object, 'files' is the list of files that are modified,
added or removed for this change, an empty list means that no files were
modified, so no reconfiguration is launched.
A source driver should fill 'change.files' info if such data are available
or can be easily fetched, otherwise 'change.files' is kept to the initialized
'None' value to trigger a reconfiguration because we can't know if it touches
the configuration or not.

For Gerrit, it's possible to get the list of files associated to a commit
using REST API 'projects/.../commits/.../files', only works if http session
is available.

on branch_created event:
-ssh: only change info can be queried, commit files can't.
 'change.files' is kept to None, triggering a tenant reconfiguration
-http(+ssh): files are queried using http REST API, change.files is a valid
 list

Change-Id: I5628039758ae3357c0c54502f01cf5e2a58caf57