Tenant reconfiguration: Add tests
Add tests before introducing tenant reconfiguration changes.
This highlights the current behaviour for multiple scenarios.
Expected values are set to pass tests.
It will be easier to spot differences introduced by the following
changes.

Gerrit stream-events not enabled
--------------------------------
A fix to the way change-merged events are generated is required.
The same change is generating multiple change-merged events due to
the way poll_delay is used in tests. It was causing unexpected tenant
reconfigurations.

One way to solve this could be to remove the overlap introduced by
the extra seconds when querying merged changes, so a merged change
appears only in a single query.
The other one, applied here, is to create a small cache to remember
which changes have already generated a change-merged event. The cache
is cleared when no merged change has been seen for a query, which should
happen most of the time with the real class/object.

As "change-merged" and "ref-updated" are both generated py polling threads
and ref-updated's branch_updated is set to True by GitWatcher,
"ref-updated" events are curently disabled in tests when modifying the
repository with a gerrit change.

Change-Id: I53a532dff59334017775747d8c0a4c83f3a19c45