Update patch set 2
Patch Set 2:

Hi, I just did some investigation on why the scrolling is not working as expected and the problem seems to be the drawer component which we use to display config errors. When I remove the drawer component (which is kind of a wrapper around the page content) the scrolling works as expected. I'm not sure if that is the expected behaviour of the drawer or if I just used it the wrong way. I will check if I can find a fix for that.

The main drawback I see with that fix here (apart from the scroll bars) is that the navigation bar is not fixed anymore.

Patch-set: 2
Reviewer: Gerrit User 27952 <27952@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0