PF4: Fix scrolling behaviour when using keyboard
Introducing PF4 broke the scrolling behaviour when using the keyboard
(e.g. spacebar, pageup/down, arrow keys) and most probably also screen
readers (which I don't use so I can't say for sure).

I'm still not entirely sure about the exact root cause, but it seems to
be related to the Drawer component and/or the way we use it (by wrapping
the entire page content). However, taking a look at the PF4 design
guidelines [1] and the linked components [2] this seems to be the way to
go.

When using the Drawer component it looks like the focus gets lost on the
page and we are not able to regain it (e.g. by simply clicking somewhere
on the page). The only solution to get the focus back is by clicking on
an input field, checkbox or some other "clickable" element. If such an
element is in focus, everything works fine.

While investigating this "focus behaviour" I noticed that the main
content of the page has a tabindex="-1". This value is used by PF4 by
default [3], but somehow seems to conflict with the Drawer component.

Thus by unsetting this value we can fix the scrolling behaviour while
using the Drawer component. However, unsetting this value without using
the Drawer component would also break the scrolling behaviour, so it
seems like this must be aligned somehow.

As a side note: In the last weeks a few fixes for the Drawer component
came in, some of them where related to focus , e.g. [4] (which allows
focusing the drawer content when open). So, I think we should observe
the scrolling behaviour, especially when updating to a newer PF4
release. On the other hand, it could still be a bug in the component
and there might be another fix for this.

[1] https://www.patternfly.org/v4/design-guidelines/usage-and-behavior/notification-drawer#placement-in-a-page
[2] https://www.patternfly.org/v4/design-guidelines/usage-and-behavior/notification-drawer#react
[3] https://www.patternfly.org/v4/documentation/react/components/page#props
[4] https://github.com/patternfly/patternfly-react/pull/4328

Change-Id: I835c56bb6d7bd84df626acac1d0f8bec16cf58a8