Update patch set 1
Patch Set 1: Code-Review-1

Thanks for looking into it ... This is definitely closer but I still see a problem maybe it's worth looking into.

My exact path is to click on zuul-build-dashboard-opendev -> site preview -> openstack | status

When that loads I initially can not use the keyboard to navigate.  I have to click inside the status page before it starts to work.

After more investigation, I found I can toggle this behaviour by clicking in the top, black navigation bar -- that stops the keyboard being able to scroll the panel -- then if i click in the panel again I can use the keys to scroll.

This doesn't seem to be focus related.  I put in the debug console

 document.addEventListener('keyup', function() { console.log(document.activeElement) })

and this constantly returns "<body>" (unless I click on something like a text input box) ... I'm not sure if that's the "null" value or that's really where the focus is?  This makes me wonder if even we did something like give the panel focus on mount if it will do anything anyway.  Perhaps react has some internal notion of the focus where it's sending keystrokes?

Patch-set: 1