Give the focus to a <div> overlaying the main content
This gives the focus to the main content when the page loads, so
navigation keys "just work".  Without this you have to click in the
content page before it will scroll.

There is still a bifurcation between the content and the navigation
bar; if you click in the nav bar anywhere, keystrokes stop working to
scroll the main page again and you'll still have to click in the panel
to get it responding again.

Change-Id: I95d593cc25b4848b1c690f91afc841d897a3b9cd