Update patch set 1
Patch Set 1:

To try and be a bit more constructive; https://review.opendev.org/743440 overlays a div on the main panel that is given focus on render.  this seems to work, in so much as you can scroll with keys as soon as you load the page, but I'm sure it is nothing but a horrible hack.  I feel like it is a data point as to what's going on though ...

Another thing to maybe consider -- if the floating nav bar is causing problems; is it really a good use of screen real estate anyway?  when i'm say, scrolling a long log file, i'd probably prefer to have those pixels showing log that a menu I'm not using.  I've found it hard to find an analogue that keeps a full floating nav bar.  

Sites like google results and github seem to do a thing where as you scroll down, it shrinks into a smaller floating header with just some relevant info (in google's case a search bar, github gives the issue/pull request title and what not).  Other sites like patternfly itself let the navbar scroll off, even on their long documentation pages.  i don't feel strongly but might be worth thinking about.

Patch-set: 1