Update patch set 1
Patch Set 1:

Now that I've read all of the answers :D
I did also try something like the ref, but not on a wrapper div, but just a (empty) span on top of the page. Also with setting focus during componentDidMount(). Just - if I remember correctly - that alone did not do the trick as I was still not able to focus the page itself (just by clicking on buttons or input fields). That would speak for a combination of both, but I agree that we should not try to fix something we don't understand with a lot of workarounds. That's why I came up with that solution in the first place as it looked like the less intrusive way to get most of the expected behaviour back ;-)

Regarding the fixed navbar: That's a good point and I did also thought if the height: 100% setting on the whole page is the reason for this. But I'm very sure that when I looked at the patternfly examples, demos, ... they all used a fixed navbar. That's why I did that in the first place :D I personally don't have a strong opinion on this and I also think that - if we want to go along with other modern websites we should go for the minimalistic navbar when scrolling down. But I think that is a topic for itself. So I would also be fine with not having a fixed navbar but scrolling working as expected.

Patch-set: 1