Update patch set 1
Patch Set 1:

> https://www.patternfly.org/v4/documentation/react/demos/masterdetaildemo/master-detail-full-page

I agree that this exhibits the same behaviour as we're seeing.  I think we agree it's a bug to behave like this?  

That demo seems to be a pretty good way to file an actionable thing upstream, since it happens on their own code (easier than trying to make someone parse a whole other website).  Could we file a bug?  then personally I'd be more happy having work-arounds in our code with a link to that bug because it seems like we've at least got the possibility of having it work better, even if it doesn't right now :)

Patch-set: 1