Update patch set 1
Patch Set 1:

I've asked about this in their slack channel and the reply was that setting the mainTabIndex is "likely the correct fix":

> This is likely the correct fix. We added that mainTabIndex to Page in
> our breaking change release for a11y reasons. Afterwards we added the
> option to set it to null to fix tab behavior in some apps. I can't
> find the original issue to add it, but now I'm curious why we included
> it in the first place

I also raised a follow-up question this morning about how to set focus direcly and if that's the intended behaviour or a bug. I think we should wait for the response before filing a bug report.

Patch-set: 1