Update patch set 1
Patch Set 1:

Based on the replies I got in the slack channel, we are on the right track with this solution and this whole "focus" topic seems very complex especially when using the Drawer to wrap the whole page.

There is also a fix in the newest PF4 release that allows focusing a different element when the Drawer is closed (which is also not working by now) [1] and [2]. So I assume that we have to deal with setting the focus by ourselves anyway, so using this fix together with a ref to focus on in componentDidMount() looks like the proper solution for now.

I've still filed a bug report for this behaviour [3], as this should IMHO work "out-of-the-box" for simple page layouts at least.

[1]: https://github.com/patternfly/patternfly-react/pull/4328
[2]: https://github.com/patternfly/patternfly-react/pull/4508
[3]: https://github.com/patternfly/patternfly-react/issues/4624

Patch-set: 1