Update patch set 1
Patch Set 1: -Code-Review

Thanks for investigating.  I'd probably like to see that bug referenced inline to any workarounds if we have another revision.  

If I've understood we need this *and* the focus change on mount to some element in the main panel?  should we squash that together?

Patch-set: 1
Label: Code-Review=0