PF4: Use a ref to focus main content on initial page load
This should allow us to directly scroll via keyboard on initial page
load. For more information on the background of this topic, take a
look at the discussion in the parent change [1].

However, this solution has one drawback: As we are directly setting
focus to an element in the main content of the page, tabbing through the
page will directly start there and skip the navigation bar / page header
with all its elements.

I've filed a bug report on Github [2] about this whole "scrolling and
focus" topic and updated it with my latest findings, so we can refer to
this one (and I don't need to write everything in multiple places :D ).

So my assumption is now that this whole topic is more related to the
fixed navbar than the Drawer component. I've compared the Zuul UI with
the official PF4 website, Github and Google in that regard and all three
of those are using a non-fixed navbar while having the focus set to the
body element (and some element that is contained in the main content).

As the question was raised if we could also use a non-fixed navbar, I
asked this question in the Patternfly slack channel as I couldn't find
any information about this in the docs. For more information, please
take a look at the Github issue.

[1]: https://review.opendev.org/#/c/743239/
[2]: https://github.com/patternfly/patternfly-react/issues/4624

Change-Id: I9b9f1cc0638f4d452e8cc9c299f418a61e8679ab