Update patch set 2
Patch Set 2:

I've added another change on top which implements the "ref-solution". It's more or less the same like you did just with a little less code. I've also linked the Github issue there.

However, I've discovered this morning that this is still not a proper solution and updated the Github issue accordingly. So we might use this one for reference.

Patch-set: 2