ConfigItem*Exception : add failure context
These various exceptions are raised as certain parts of the
configuration are parsed.  Although they give you a lot of information
about the issue, they don't give you any context for what part of your
configuration caused the error to be raised.

They are all working on a particular item of the configuration that I
think it would be very useful to show in the error output.

This was inspired by me putting playbooks in the .zuul.d directory,
where they were read as config files.  In hindsight it was obvious,
but it took me a while to figure out what was going on.  With this it
would have reported something like:

  Zuul encountered a syntax error while parsing its configuration in
  the repo org/project1 on branch master.  The error was:

  Configuration item has more than one key.  Each zuul.yaml
  configuration file must be a list of dictionaries with a
  single key, for example:

  - job:
      name: foo
  ... blah ...

  The incorrect values are:

  hosts: all
  tasks:
  - debug:
      msg: blah
    name: foo

which would have clued me in immediately.

This truncates the message if it is too long; which might be the case
if dumped a playbook or foreign config file.  However I think
practically most errors are typo level on jobs/templates/etc. that
will show nicely.

The YAML version is updated to support non-sorted dumping of the keys.
This is particularly helpful with Python 3.7+ and default ordered
dicts.

Change-Id: I4851a5a796fa452a023e0e6193fc724ae0967a44