Update patch set 2
Patch Set 2:

Just wanted to sum this topic up with the latest information I got from the patternfly devs:

Currently there is no perfect solution for this topic. The one implemented in this (and the follow-up) change allows the scrolling, but users will skip the navbar while tabbing through the page. But IMHO this should be ok for now.

A different approach (at least I'm very sure that this would help) would be a floating navbar. However, using a floating navbar is not supported by PF4 by now. Although they are using one on the official PF4 website, it's more like a hack to achieve it. I could use a similar hack in our UI, but this would break the Drawer placement and (kind of) its animation. As they don't use any Drawer on the PF4 website it's not a problem there. Based on this, it could be that a floating navbar is coming but not for sure.

So, all in all I would go for the currently proposed solution and try to deal with the floating navbar again after some time. Since Ian raised this topic, I think it would be really cool to have something like Google or Github (not a fixed navbar, but a floating one that kind of minimizes when scrolling down).

Patch-set: 2