Remove unneeded api requests when commenting in github
Currently when commenting on a pr in github we first get the repo
object and then the issue object. After that we call create_comment on
this object. This issues two unnecessary api calls to get project and
issue. Instead directly build the correct url and do the post manually
to comment.

Change-Id: I9afb19e4918109e98882aa9d5b716d34c65936f2