Update patch set 2
Patch Set 2:

> If you put in a very large playbook, would it output the entire playbook here?

I think it depends on the error; so if it's finding something at the very top level of the playbook then I guess yes it shows that, but if it's already deeper into the config it will only show the element it's working on.

> Maybe we could include +/- 5 raw lines.

That might work, or somehow not including children elements and just showing the actual thing it's working on?  I'll investigate both as I think it would be helpful to give *some* context, even minimal. 

> However, I think a better approach would be to rely on inline comments.

I agree this is the gold standard.  It seemed to me we'd already parsed and lost that information at this point; especially since it may have merged changes together.  If you already have thoughts about this and have some tips on how we could keep the association I'm all ears :)

Patch-set: 2